+++ gcc/lto-compress.h
@@ -22,14 +22,14 @@ along with GCC; see the file COPYING3.  If not see
 struct lto_compression_stream;
 
 extern struct lto_compression_stream
-  *lto_start_compression (void (*callback) (const char *, size_t, void *),
+  *lto_start_compression (void (*callback) (const char *, unsigned, void *),
 			  void *opaque);
 extern void lto_compress_block (struct lto_compression_stream *stream,
 				const char *base, size_t num_chars);
 extern void lto_end_compression (struct lto_compression_stream *stream);
 
 extern struct lto_compression_stream
-  *lto_start_uncompression (void (*callback) (const char *, size_t, void *),
+  *lto_start_uncompression (void (*callback) (const char *, unsigned, void *),
 			    void *opaque);
 extern void lto_uncompress_block (struct lto_compression_stream *stream,
 				  const char *base, size_t num_chars);
